:truck: Move Service-wide Auth Manager to shared
