+++ src/shared/AuthorizationManager.ts
